sandbox/requirements.txt
torch
open-clip-torch
opencv-python
Pillow
numpy
mlx-vlm
pygetwindow
mss
pyobjc
